+++ tensorflow/lite/micro/kernels/BUILD
@@ -249,6 +249,7 @@ tflm_kernel_cc_library(
         "reduce.cc",
         "reduce_common.cc",
         "reshape.cc",
+        "reshape_common.cc",
         "resize_bilinear.cc",
         "resize_nearest_neighbor.cc",
         "round.cc",
@@ -299,6 +300,7 @@ tflm_kernel_cc_library(
         "prelu.h",
         "quantize.h",
         "reduce.h",
+        "reshape.h",
         "softmax.h",
         "sub.h",
         "svdf.h",
