+++ tensorflow/lite/micro/kernels/reshape.cc
@@ -13,6 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
+#include "tensorflow/lite/micro/kernels/reshape.h"
+
 #include <cstring>
 
 #include "tensorflow/lite/c/builtin_op_data.h"
@@ -29,74 +31,11 @@ namespace ops {
 namespace micro {
 namespace reshape {
 
-constexpr int kInputTensor = 0;
-constexpr int kOutputTensor = 0;
-
-TfLiteStatus ReshapeOutput(TfLiteContext* context, TfLiteNode* node) {
-  MicroContext* micro_context = GetMicroContext(context);
-
-  TfLiteTensor* input =
-      micro_context->AllocateTempInputTensor(node, kInputTensor);
-  TF_LITE_ENSURE(context, input != nullptr);
-  TfLiteTensor* output =
-      micro_context->AllocateTempOutputTensor(node, kOutputTensor);
-  TF_LITE_ENSURE(context, output != nullptr);
-  // Tensorflow's Reshape allows one of the shape components to have the
-  // special -1 value, meaning it will be calculated automatically based on the
-  // input. Here we calculate what that dimension should be so that the number
-  // of output elements in the same as the number of input elements.
-  int num_input_elements = NumElements(input);
-  TfLiteIntArray* output_shape = output->dims;
-
-  if (NumInputs(node) == 1 &&  // Legacy scalar supported with params.
-      output_shape->size == 1 && output_shape->data[0] == 0) {
-    // Legacy tflite models use a shape parameter of [0] to indicate scalars,
-    // so adjust accordingly. TODO(b/111614235): Allow zero-sized buffers during
-    // toco conversion.
-    output_shape->size = 0;
-  }
-
-  int num_output_elements = 1;
-  int stretch_dim = -1;
-  for (int i = 0; i < output_shape->size; ++i) {
-    int value = output_shape->data[i];
-    if (value == -1) {
-      TF_LITE_ENSURE_EQ(context, stretch_dim, -1);
-      stretch_dim = i;
-    } else {
-      num_output_elements *= value;
-    }
-  }
-  if (stretch_dim != -1) {
-    TfLiteEvalTensor* output_eval =
-        tflite::micro::GetEvalOutput(context, node, kOutputTensor);
-    TF_LITE_ENSURE_STATUS(tflite::micro::CreateWritableTensorDimsWithCopy(
-        context, output, output_eval));
-    output_shape = output->dims;  // output tensor dims were moved
-    output_shape->data[stretch_dim] = num_input_elements / num_output_elements;
-    num_output_elements *= output_shape->data[stretch_dim];
-  }
-
-  TF_LITE_ENSURE_TYPES_EQ(context, input->type, output->type);
-  TF_LITE_ENSURE_EQ(context, num_input_elements, num_output_elements);
-
-  micro_context->DeallocateTempTfLiteTensor(input);
-  micro_context->DeallocateTempTfLiteTensor(output);
-  return kTfLiteOk;
-}
-
-TfLiteStatus Prepare(TfLiteContext* context, TfLiteNode* node) {
-  TF_LITE_ENSURE(context, NumInputs(node) == 1 || NumInputs(node) == 2);
-  TF_LITE_ENSURE_EQ(context, NumOutputs(node), 1);
-  TF_LITE_ENSURE_EQ(context, ReshapeOutput(context, node), kTfLiteOk);
-  return kTfLiteOk;
-}
-
 TfLiteStatus Eval(TfLiteContext* context, TfLiteNode* node) {
   const TfLiteEvalTensor* input =
-      tflite::micro::GetEvalInput(context, node, kInputTensor);
+      tflite::micro::GetEvalInput(context, node, kReshapeInputTensor);
   TfLiteEvalTensor* output =
-      tflite::micro::GetEvalOutput(context, node, kOutputTensor);
+      tflite::micro::GetEvalOutput(context, node, kReshapeOutputTensor);
 
   // TODO(b/162522304): storing input bytes in OpData increases some models
   // significantly, possibly due to alignment issues.
@@ -115,7 +54,8 @@ TfLiteStatus Eval(TfLiteContext* context, TfLiteNode* node) {
 }  // namespace reshape
 
 TFLMRegistration Register_RESHAPE() {
-  return tflite::micro::RegisterOp(nullptr, reshape::Prepare, reshape::Eval);
+  return tflite::micro::RegisterOp(nullptr, reshape::PrepareReshapeReference,
+                                   reshape::Eval);
 }
 
 }  // namespace micro
