+++ tensorflow/lite/micro/kernels/reshape_test.cc
@@ -1,4 +1,4 @@
-/* Copyright 2017 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2023 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -62,6 +62,7 @@ void ValidateReshapeGoldens(TfLiteTensor* tensors, int tensors_size,
     TF_LITE_MICRO_EXPECT_EQ(expected_dims[i], output_tensor->dims->data[i]);
   }
 }
+
 template <typename T>
 void TestReshapeWithShape(TfLiteTensor* input_tensor,
                           TfLiteTensor* shape_tensor,
@@ -172,9 +173,7 @@ TF_LITE_MICRO_TEST(ReshapeWithMismatchedDimensionsShouldFail) {
       golden_output, golden_output_len, golden_dims, golden_dims_len, true);
 }
 
-// TODO(b/237407410): Re-enable for Vision P6 when the issue is resolved
-#if !defined(VISION_P6)
-TF_LITE_MICRO_TEST(ReshapeWithTooManyDimensionsShouldFail) {
+TF_LITE_MICRO_TEST(ReshapeWithManyDimensionsShouldSucceed) {
   float output_data[32];
   int input_dims[] = {9, 1, 1, 2, 1, 1, 1, 1, 1, 1};
   const float input[] = {3, 2};
@@ -189,7 +188,6 @@ TF_LITE_MICRO_TEST(ReshapeWithTooManyDimensionsShouldFail) {
       input_dims, input, shape_dims, shape_int32, output_dims, output_data,
       golden_output, golden_output_len, golden_dims, golden_dims_len, false);
 }
-#endif
 
 TF_LITE_MICRO_TEST(ReshapeWithTooManySpecialDimensionsShouldFail) {
   float output_data[32];
